Move chats and contacts file out of omnichannel directory
