+++ client/omnichannel/contacts/contextualBar/ContactForm.js
@@ -3,20 +3,20 @@ import { Field, TextInput, ButtonGroup, Button, Box } from '@rocket.chat/fuselag
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 import { useSubscription } from 'use-subscription';
 
-import { useTranslation } from '../../../../contexts/TranslationContext';
-import VerticalBar from '../../../../components/VerticalBar';
-import { useForm } from '../../../../hooks/useForm';
-import { isEmail } from '../../../../../app/utils';
-import { useComponentDidUpdate } from '../../../../hooks/useComponentDidUpdate';
-import { useEndpointAction } from '../../../../hooks/useEndpointAction';
-import { useToastMessageDispatch } from '../../../../contexts/ToastMessagesContext';
-import { useEndpointData } from '../../../../hooks/useEndpointData';
-import { FormSkeleton } from '../../Skeleton';
-import CustomFieldsForm from '../../../../components/CustomFieldsForm';
-import { hasAtLeastOnePermission } from '../../../../../app/authorization';
-import { AsyncStatePhase } from '../../../../hooks/useAsyncState';
-import { formsSubscription } from '../../../../views/omnichannel/additionalForms';
-import { createToken } from '../../../../components/helpers';
+import { useTranslation } from '../../../contexts/TranslationContext';
+import VerticalBar from '../../../components/VerticalBar';
+import { useForm } from '../../../hooks/useForm';
+import { isEmail } from '../../../../app/utils';
+import { useComponentDidUpdate } from '../../../hooks/useComponentDidUpdate';
+import { useEndpointAction } from '../../../hooks/useEndpointAction';
+import { useToastMessageDispatch } from '../../../contexts/ToastMessagesContext';
+import { useEndpointData } from '../../../hooks/useEndpointData';
+import { FormSkeleton } from '../../directory/Skeleton';
+import CustomFieldsForm from '../../../components/CustomFieldsForm';
+import { hasAtLeastOnePermission } from '../../../../app/authorization';
+import { AsyncStatePhase } from '../../../hooks/useAsyncState';
+import { formsSubscription } from '../../../views/omnichannel/additionalForms';
+import { createToken } from '../../../components/helpers';
 
 const initialValues = {
 	token: '',
