+++ client/views/room/lib/Toolbox/defaultActions.ts
@@ -28,7 +28,7 @@ addAction('contact-profile', {
 	id: 'contact-profile',
 	title: 'Contact_Profile',
 	icon: 'user',
-	template: lazy(() => import('../../../../omnichannel/directory/contacts/contextualBar')),
+	template: lazy(() => import('../../../../omnichannel/contacts/contextualBar')),
 	order: 5,
 });
